Rename metadata to extra_info
metadata is a reserved keyname, and metadatas sounds dumb, so extra_info
it is.

I also attempted to get has_many => belongs_to relationship going
between days/posts using date as the primary key, but that caused all
kinds of problems. Seems like theres a moped bug when using a Date type
as the primary key.

Also set a `default_extra_info` field on the viewer, so now we don't
need to include a YAML file and we can have better support for multiple
users. I should probably get to work on making things user specific.