Fight the power man!
I'm not using SQL anymore and mongo doesn't have any kind of 255
character restriction on strings. The 255 there has always been a soft
limit anyways, and was just there to tell me to keep it short.

But I've realized 255 characters is often not enough for a brief summary
of the day, and I had to write more and more posts just to summarize my
day. Writing posts is a great thing, but it dilutes them when they
become more of an everyday thing.

500 characters is a good place to put a soft boundary.