Stop making scheduled tasks for GithubEvents
Instead of having scheduled events that rely on heroku scheduled tasks
calling a rake job, we can just have the job enqueue itself, so it just
needs to be called once, and then it reschedules itself with SideKiq.

Eventually, I might make a more generalized function that calls itself
periodically and checks which jobs needs to be run.