Clean up Photo uploading and phase out Filepicker
Filepicker got rid of their free plan, so switch over to just
uploading straight to S3, which is a better way to do it anyways.

This gets rid of the user uploading of photos, and it is all now done
through Dropbox.

A future commit will add updating of photos to the daily async tasks